# Conflicts:
#	code/app/src/main/java/com/example/project007/TrailsActivity.java
#	code/app/src/main/res/values/strings.xml

+++ code/app/src/main/java/com/example/project007/TrailsActivity.java
@@ -13,6 +13,7 @@ import android.os.Looper;
 import android.provider.Settings;
 import android.util.Log;
 import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
@@ -27,7 +28,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
@@ -37,17 +37,10 @@ import com.google.android.gms.location.LocationCallback;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationResult;
 import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.EventListener;
@@ -56,8 +49,6 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
-
-import org.w3c.dom.Text;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -89,6 +80,7 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
 
     String description;
     String title;
+
 
     android.location.Location currrentLocation;
     FusedLocationProviderClient fusedLocationProviderClient;
@@ -162,10 +154,10 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
             case "Measurement":
                 imageView.setImageResource(R.drawable.m);
                 break;
-            case "Count":
+            case "Count-base":
                 imageView.setImageResource(R.drawable.c);
                 break;
-            case "IntCount":
+            case "Non-negative":
                 imageView.setImageResource(R.drawable.n);
                 break;
         }
@@ -184,11 +176,10 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         if (experiment.getUserId().equals(DatabaseController.getUserId())) {
             owner.setVisibility(View.VISIBLE);
         }
+        //set flag to hide menu option
 
-        final FloatingActionButton addButton = findViewById(R.id.experimentBtn);
-        if (!experiment.isCondition()) {
-            addButton.setVisibility(View.INVISIBLE);
-        }
+        /*final FloatingActionButton addButton = findViewById(R.id.experimentBtn);*/
+
 
         trail_List = findViewById(R.id.trail_list);
 
@@ -282,15 +273,15 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         //add button is where we specify the different experiment trails
         //once firestrore ready this part will get type from database
         if (type.equals("Binomial")) {
-            addButton.setOnClickListener(new View.OnClickListener() {
+            /*addButton.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
                     //initialize fragment
                     addBinoTrailFragment = new AddBinoTrailFragment();
-                    getSupportFragmentManager().beginTransaction().replace(R.id.data_container, addBinoTrailFragment).addToBackStack(null).commit();
+                    getSupportFragmentManager().beginTransaction().replace(R.id.data_container2, addBinoTrailFragment).addToBackStack(null).commit();
                     //addButton.setVisibility(View.INVISIBLE);
                 }
-            });
+            });*/
 
             trail_List.setOnItemClickListener(new AdapterView.OnItemClickListener() {
                 @Override
@@ -320,14 +311,14 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
             });
 
         } else {
-            addButton.setOnClickListener(new View.OnClickListener() {
+            /*addButton.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
                     //initialize fragment
                     addNnCBTrailFragment = new AddNnCBTrailFragment();
-                    getSupportFragmentManager().beginTransaction().replace(R.id.data_container, addNnCBTrailFragment).addToBackStack(null).commit();
+                    getSupportFragmentManager().beginTransaction().replace(R.id.data_container2, addNnCBTrailFragment).addToBackStack(null).commit();
                 }
-            });
+            });*/
             trail_List.setOnItemClickListener(new AdapterView.OnItemClickListener() {
                 @Override
                 public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
@@ -359,30 +350,39 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         }
 
 
+
         //longClick action for delete data
         trail_List.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
             @Override
             public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
                 //Delete event
-                Trails newtrail = trail_Adapter.getItem(position);
-                String idString = newtrail.getID().toString();
+                String databaseUserId = DatabaseController.getUserId();
+                String experimentId = experiment.getUserId();
+                if (databaseUserId.matches(experimentId)){
+                    Trails newtrail = trail_Adapter.getItem(position);
+                    String idString = newtrail.getID().toString();
 
-                ArrayList<String> temp =  experiment.getTrailsId();
-                for (int i = 0; i < temp.size(); i++) {
-                    if (temp.get(i).equals(idString)) {
-                        temp.remove(i);
+                    ArrayList<String> temp =  experiment.getTrailsId();
+                    for (int i = 0; i < temp.size(); i++) {
+                        if (temp.get(i).equals(idString)) {
+                            temp.remove(i);
+                        }
                     }
-                }
-                DatabaseController.setExperimentTrails(experiment.getId().toString(),temp,experiment.getSubscriptionId());
+                    DatabaseController.setExperimentTrails(experiment.getId().toString(),temp,experiment.getSubscriptionId());
 
-                trail_Adapter.notifyDataSetChanged();
+                    trail_Adapter.notifyDataSetChanged();
 
-                boolean deleteResult = TrailsDatabaseController.delete_Trails("Trails", newtrail);
-                if (deleteResult) {
-                    Toast.makeText(getApplicationContext(), "Delete Succeed", Toast.LENGTH_SHORT).show();
-                } else {
-                    Toast.makeText(getApplicationContext(), "Delete Failed", Toast.LENGTH_SHORT).show();
+                    boolean deleteResult = TrailsDatabaseController.delete_Trails("Trails", newtrail);
+                    if (deleteResult) {
+                        Toast.makeText(getApplicationContext(), "Delete Failed", Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(getApplicationContext(), "Delete Succeed", Toast.LENGTH_SHORT).show();
+                    }
+                }else{
+                    Toast.makeText(getApplicationContext(), "You don't have access to delete this trail", Toast.LENGTH_SHORT).show();
                 }
+
+                //revert logic
                 return false;
             }
         });
@@ -390,6 +390,7 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         //https://stackoverflow.com/questions/4834750/how-to-get-the-selected-item-from-listview
         //from xandy's answer Jan 29 '11 at 2:57*/
     }
+
 
     private void chooseIgnore(int id, Trails newtrail, String remoteUserId, String localUserId) {
         final FirebaseFirestore db;
@@ -522,10 +523,11 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         }
         DatabaseController.setExperimentTrails(experiment.getId().toString(), valueList,subscriptionList );
         if (addResult){
-            Toast.makeText(getApplicationContext(), "Add Succeed", Toast.LENGTH_SHORT).show();
-        }else{
             Toast.makeText(getApplicationContext(), "Add Failed", Toast.LENGTH_SHORT).show();
+        }else{
+            Toast.makeText(getApplicationContext(), "Add Succeed", Toast.LENGTH_SHORT).show();
         }
+        //revert logic
     }
 
     @Override
@@ -554,7 +556,14 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         }
     }
 
-
+    @Override
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        if (!experiment.isCondition()) {
+            menu.findItem(R.id.action_add).setVisible(false);
+        }
+        super.onPrepareOptionsMenu(menu);
+        return true;
+    }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -562,6 +571,7 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
         getMenuInflater().inflate(R.menu.menu_trails_activity, menu);
         return true;
     }
+
 
     //YO!!! This is where you implement those fragments under the if
 
@@ -613,9 +623,18 @@ public class TrailsActivity extends AppCompatActivity implements AddBinoTrailFra
             return true;
         }else if (id == R.id.HelpOpt){
             //tips for user
-            Toast.makeText(getApplicationContext(),"Welcome! Please note: Long Click item for deleting Short Click item for editing",Toast.LENGTH_SHORT).show();
+            Toast.makeText(getApplicationContext(),"Welcome! Please note: Long Click item for item deleting~",Toast.LENGTH_SHORT).show();
             //Toast.makeText(getApplicationContext(),trails_DataList.toString(),Toast.LENGTH_SHORT).show();
             return true;
+        }else if(id == R.id.action_add){
+            if (type.equals("Binomial")) {
+                addBinoTrailFragment = new AddBinoTrailFragment();
+                getSupportFragmentManager().beginTransaction().replace(R.id.data_container2, addBinoTrailFragment).addToBackStack(null).commit();
+            }else{
+                addNnCBTrailFragment = new AddNnCBTrailFragment();
+                getSupportFragmentManager().beginTransaction().replace(R.id.data_container2, addNnCBTrailFragment).addToBackStack(null).commit();
+            }
+
         }
 
         return super.onOptionsItemSelected(item);
